requirements.txt
jose==1.0.0
Flask==1.1.1
alembic==1.4.2
Flask_SQLAlchemy==2.4.1
Flask_Migrate==2.5.3
Flask_Cors==3.0.8
six==1.11.0
requests==2.20.1
Werkzeug==0.15.4
Authlib==0.14.1
pytest==5.4.1
SQLAlchemy==1.3.15
python-dotenv==0.12.0
